Implement a farm_trait module
Provide a field collection `field_farm_trait` which can be
used for tracking arbitrary traits associated with crops/
varieties, livestock animal types, or specific assets.

Traits are modeled as a key-value pair where the key comes
from a new taxonomy `farm_trait_types` and the value is an
unconstrained text field. The trait types taxonomy includes
a description - which could be used for documenting how the
trait is measured - and a reference ID which could be used
for linking to a canonical ontology.
e.g http://purl.obolibrary.org/obo/TO_0002734