Run all PHPUnit tests in the /opt/drupal/web/profiles/farm directory, instead of using --group farm.
The --group flag fails with a fatal error if there are any tests (even
not in the requested group) that extend from a PHP class that
doesn't exist. This is true for some modules like fraction and
email_registration which have tests that extend from tests in
other modules that we do not include in our composer.json.