Save quantities as a pseudo field in the farm_migrate_log migration group.
This is necessary so that individual log types can use pseudo fields in
the "quantity" process if they override it themselves.

When a migration group defines shared_configuration processes that are
overwritten by a migration, the processes are still executed in the
order defined by the migration group. This prevents an individual
migration from using its own pseudo fields for a process that it
overrides.