Issue #3239420: Use composer-merge-plugin to manage project composer.json dependencies
This enables more of the dependencies and configuration that need to be
present in the root project composer.json to be managed in the farmOS
repository itself, including drupal/core-composer-scaffold and
drupal/core-dev, which should match the drupal/core version defined in
the farmOS composer.json.

Previously these were declared in the project composer.json template
in https://github.com/farmOS/composer-project. This configuraton is
used as a template when a new farmOS project is created via
`composer create-project farmos/project`, but after initial setup it
is maintained by the end user. This means that updates to the project
composer.json generally requires the end-user to make changes
manually. This would be required every time farmOS requires a minor
version update of Drupal core.

So instead, we can manage those update in this new composer.project.json,
and use wikimedia/composer-merge-plugin to merge that into the project
composer.json. This allows the project composer.json to be much simpler,
and therefore will require less frequent manual updates.