Reference non-existent farmos/farmos:2.x.y Docker tag in production hosting docs/examples.
By doing this, we will be forcing the end-user to follow best
practice by making an intentional decision about the version
they will use. We do not want farmos/farmos:2.x to be referenced
by default in the docs, otherwise we will be directing folks to
use the development branch, which may not be stable, and may
change if they update the image via `docker pull`. We also don't
want to recommend farmos/farmos:latest, for a similar reason:
when updating between versions it is important to manually run
update.php. By forcing intentional version use and updates, we
encourage best practice for updates as well.