Adapt csv export logic to support a type filter allowing single or multiple values. #584
The current logic is expecting the exposed type filter to always be an array. This causes
csv exports to break when exporting from the all quantities page because the quantity
type filter does not allow multiple values. When no quantity type is provided the type
filter is set to 'All'.