Remove default_measure configuration from quantity type definitions #612
This was originally added so that the Price quantity type could set its
default measure to "value" all the time. This was back when price
quantities were part of farmOS core. It was then moved to the
farm_ledger contrib module, and is being refactored such that it will no
longer need to make use of this default_measure configuration.