Improve asset bundle logic in ID tag widget #699
This fixes a bug where assets cannot be used in an inline entity form because
the ID tag widget assumes it is being used in the context of an entity
form_display. Inline entity forms and likely other entity forms may not use
the form_display key in form storage so using the entity from the
FieldItemList is more robust.