Fix issue with build-farmOS.sh tagged releases.
Previously ${FARMOS_COMPOSER_VERSION} was not being set for
tagged releases, causing `composer require farmos/farmos`
to run without a version constraint. This worked for 2.x
releases because it would default to the latest stable release.

This does not work for 3.0.0-betaX releases, because they are
not considered "stable" by Composer, so it falls back on the
latest stable 2.x release (2.2.2).