Remove farm.settings.farm_info menu link.
This is currently the only sub-menu link provided for
settings pages, which is weird. This is a quick fix which
now forces users to click "Settings" to see all the sub-tabs,
instead of thinking there is only one by looking at the
toolbar menu. In a follow-up we may consider providing a
deriver that automatically generates menu links for each
sub-tab so they all show in the toolbar menu. It would be
burdensome to require every module that provides a settings
tab to also add a menu link.