Fix saving inventory quick form configuration with existing units term #803
The entity_autocomplete $form_state value differs depending on
whether it is referencing an existing term or creating a new
term.

This fixes the submit logic in the configuration form of the
inventory quick form to handle both cases. Before this change,
it was only possible to create a new term. Referencing an
existing term would result in a PHP error:

TypeError: "Cannot access offset of type string on string"